+++ src/component/TimeBar.js
@@ -1,11 +1,7 @@
 import React from 'react'
 
-const Timebar = (props) => {
-  const { fill } = props
-
+export default function TimeBar({ fill }) {
   return (
     <rect x='0' y='0' width='86400' height='2320' fill={fill} fillOpacity='.16' />
   )
 }
-
-export default Timebar