+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add Formik as dependency
 * Add validation to Login form
 * Use Context API for handling themes
+* Refactor some code and create new components (Timebar, TaskBar)
 
 # (2018-11-05)
 * Add Login module